+++ src/data/default-crew.ts
@@ -11,4 +11,6 @@ export const defaultCrew: string[] = [
     "anaerobic",
     "random",
     "CREW_RANDOM",
+    'ghost',
+    'traitor'
 ];