+++ CHANGELOG.md
@@ -37,10 +37,16 @@ All notable changes to the "ftlXml" extension will be documented in this file.
 ## 0.2.2
 - Update readme
 
-## 0.2.3
+## 0.3.0
 - Add default text ids from misc, sectorname and tooltips
 - fix hidden arg issues
 - remove duplicate elements in autocomplete
 - allow a blueprint list to contain anything and don't warn about type
 - fix an issue where blueprint loops would break the parser
 - when a workspace folder is added or removed update the data
+
+## 0.3.1
+- Add more text id bindings for weapon text
+- add ghost and traitor as default crew names
+- add required elements for weaponBlueprints
+- add rarity autocomplete